move around, factor out the Names module
