+++ persistent/Database/Persist/Types.hs
@@ -1,5 +1,6 @@
 module Database.Persist.Types
     ( module Database.Persist.Types.Base
+    , module Database.Persist.Names
     , SomePersistField (..)
     , Update (..)
     , BackendSpecificUpdate
@@ -12,6 +13,7 @@ module Database.Persist.Types
     , OverflowNatural(..)
     ) where
 
+import Database.Persist.Names
 import Database.Persist.Types.Base
 import Database.Persist.Class.PersistField
 import Database.Persist.Class.PersistEntity
