--- persistent/Database/Persist/Types/FieldDef.hs
@@ -1,6 +1,0 @@
-module Database.Persist.Types.FieldDef
-    (
-    ) where
-
-import Database.Persist.Types.FieldDef.Internal
-
