+++ persistent/persistent.cabal
@@ -52,9 +52,7 @@ library
     exposed-modules: 
         Database.Persist
         Database.Persist.Types
-        Database.Persist.Types.Names
-        Database.Persist.Types.FieldDef
-        Database.Persist.Types.FieldDef.Internal
+        Database.Persist.Names
         Database.Persist.TH
 
         Database.Persist.Quasi
